Allow enrollment in experiments and rollouts via devtools
When we were checking for a previous enrollment, we weren't checking for
the right type of enrollment (rollout vs experiment) so if there was an
active experiment you couldn't enroll in a rollout without triggering
unenrollment of the experiment (and vice versa). As a result we also got
the name of the enrollment we were trying to enroll incorrect in the
dialog.

Fixes #101